Files/requirements.txt
mysql_connector_repackaged==0.3.1
pydantic==2.7.1
